score-calculation/requirements.txt
annotated-types==0.5.0
pika==1.3.2
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
typing_extensions==4.7.1
